+++ Sources/Utilities/ALKMessageViewModel+Extension.swift
@@ -95,6 +95,7 @@ extension ALKMessageViewModel {
     }
 
     func allButtons() -> SuggestedReplyMessage? {
+        let isActionButtonHidden = isActionButtonHidden()
         guard let payload = payloadFromMetadata() else { return nil }
         var buttons = [SuggestedReplyMessage.Suggestion]()
         for object in payload {
@@ -102,16 +103,20 @@ extension ALKMessageViewModel {
                   let action = object["action"] as? [String: Any],
                   let type = action["type"] as? String
             else { continue }
-            var buttonType = SuggestedReplyMessage.SuggestionType.normal
+            
             if type == "link" {
-                buttonType = .link
+                buttons.append(SuggestedReplyMessage.Suggestion(title: name, type: .link))
+            } else if type == "submit" {
+                buttons.append(SuggestedReplyMessage.Suggestion(title: name, type: .submit))
+            } else if !isActionButtonHidden {
+                buttons.append(SuggestedReplyMessage.Suggestion(title: name, type: .normal))
             }
-            buttons.append(SuggestedReplyMessage.Suggestion(title: name, type: buttonType))
         }
         return SuggestedReplyMessage(suggestion: buttons, message: messageDetails())
     }
 
     func suggestedReply() -> SuggestedReplyMessage? {
+        let isActionButtonHidden = isActionButtonHidden()
         guard let payload = payloadFromMetadata() else { return nil }
         var buttons = [SuggestedReplyMessage.Suggestion]()
         for object in payload {
@@ -123,13 +128,16 @@ extension ALKMessageViewModel {
     }
 
     func linkOrSubmitButton() -> SuggestedReplyMessage? {
+        let isActionButtonHidden = isActionButtonHidden()
         guard let payload = payloadFromMetadata() else { return nil }
         var buttons = [SuggestedReplyMessage.Suggestion]()
         for object in payload {
             guard let name = object["name"] as? String else { continue }
             if let type = object["type"] as? String, type == "link" {
                 buttons.append(SuggestedReplyMessage.Suggestion(title: name, type: .link))
-            } else {
+            } else if let type = object["type"] as? String, type == "submit" {
+                buttons.append(SuggestedReplyMessage.Suggestion(title: name, type: .submit))
+            } else if !isActionButtonHidden {
                 buttons.append(SuggestedReplyMessage.Suggestion(title: name))
             }
         }
@@ -156,5 +164,25 @@ extension ALKMessageViewModel {
             print("Error while decoding video template \(error.localizedDescription)")
             return nil
         }
+    }
+    
+    func isSuggestedReply() -> Bool {
+        guard self.messageType == .allButtons ||
+                self.messageType == .button ||
+                self.messageType == .quickReply else {
+            return false
+        }
+        return true
+    }
+    
+    func isActionButtonHidden() -> Bool {
+        guard UserDefaults.standard.bool(forKey: SuggestedReplyView.hidePostCTA),
+              self.isSuggestedReply(),
+              let currentMessageTime = self.createdAtTime,
+              let lastSentMessageTime = ALKConversationViewModel.lastSentMessage?.createdAtTime,
+              currentMessageTime .int64Value < lastSentMessageTime .int64Value else {
+            return false
+        }
+        return true
     }
 }