+++ Sources/ViewModels/ALKConversationViewModel.swift
@@ -49,6 +49,7 @@ open class ALKConversationViewModel: NSObject, Localizable {
 
     open var isSearch: Bool = false
     open var lastMessage : ALMessage?
+    internal static var lastSentMessage : ALMessage?
 
     // For topic based chat
     open var conversationProxy: ALConversationProxy? {
@@ -183,6 +184,9 @@ open class ALKConversationViewModel: NSObject, Localizable {
         alMessageWrapper.addALMessage(toMessageArray: message)
         alMessages.append(message)
         messageModels.append(message.messageModel)
+        if(message.isMyMessage){
+            ALKConversationViewModel.lastSentMessage = message
+        }
     }
 
     func clearViewModel() {
@@ -296,7 +300,9 @@ open class ALKConversationViewModel: NSObject, Localizable {
     open func heightForRow(indexPath: IndexPath, cellFrame _: CGRect, configuration: ALKConfiguration) -> CGFloat {
         let messageModel = messageModels[indexPath.section]
         let cacheIdentifier = (messageModel.isMyMessage ? "s-" : "r-") + messageModel.identifier
-        if let height = HeightCache.shared.getHeight(for: cacheIdentifier) {
+        let isActionButtonHidden = messageModel.isActionButtonHidden()
+        if let height = HeightCache.shared.getHeight(for: cacheIdentifier),
+           !isActionButtonHidden {
             return height
         }
         switch messageModel.messageType {
@@ -1389,6 +1395,10 @@ open class ALKConversationViewModel: NSObject, Localizable {
             if !KMConversationScreenConfiguration.staticTopMessage.isEmpty {
                 self.alMessages.insert(self.getInitialStaticFirstMessage(), at: 0)
             }
+            
+            if(ALKConversationViewModel.lastSentMessage == nil){
+                ALKConversationViewModel.lastSentMessage = self.getLastSentMessage()
+            }
 
             self.alMessageWrapper.addObject(toMessageArray: messages)
             
@@ -1397,8 +1407,7 @@ open class ALKConversationViewModel: NSObject, Localizable {
             if self.isConversationAssignedToBot() && (self.botDelayTime > 0) && !self.isOldConversation() {
                 self.showTypingIndicatorForWelcomeMessage()
             } else {
-                self.messageModels = self.modelsToBeAddedAfterDelay
-               
+                self.addMessageToMessageModel(messages: self.modelsToBeAddedAfterDelay)
             }
             self.membersInGroup { members in
                 self.groupMembers = members
@@ -1411,6 +1420,16 @@ open class ALKConversationViewModel: NSObject, Localizable {
     /*
         Since we are getting the welcome message from Api Call, we are using this method to Show Typing Delay Indicator for Welcome Messsages
      */
+    
+    func getLastSentMessage() -> ALMessage? {
+        for message in alMessages.reversed() {
+            if(message.isMyMessage){
+                return message
+            }
+        }
+        return nil
+    }
+    
     func showTypingIndicatorForWelcomeMessage() {
         if welcomeMessagePosition >= alMessages.count {
             return
@@ -1593,8 +1612,11 @@ open class ALKConversationViewModel: NSObject, Localizable {
             self.alMessages.insert(contentsOf: messages as! [ALMessage], at: 0)
 
             self.alMessageWrapper.addObject(toMessageArray: messages)
+            if(ALKConversationViewModel.lastSentMessage == nil){
+                ALKConversationViewModel.lastSentMessage = self.getLastSentMessage()
+            }
             let models = messages.map { ($0 as! ALMessage).messageModel }
-            self.messageModels.insert(contentsOf: models, at: 0)
+            self.addMessageToMessageModel(messages: models)
             if isFirstTime {
                 self.membersInGroup { members in
                     self.groupMembers = members
@@ -1605,6 +1627,27 @@ open class ALKConversationViewModel: NSObject, Localizable {
             }
         })
         self.lastMessage = alMessages.last
+    }
+    
+    func addMessageToMessageModel(messages : [ALKMessageModel]){
+        guard let lastSentMessageTime = ALKConversationViewModel.lastSentMessage?.createdAtTime,
+              UserDefaults.standard.bool(forKey: SuggestedReplyView.hidePostCTA) else {
+            self.messageModels.insert(contentsOf: messages, at: 0)
+            return
+        }
+
+        for message in messages {
+            if let currentMessageTime = message.createdAtTime {
+                if message.isMyMessage ||
+                    message.message != nil ||
+                    message.linkOrSubmitButton()?.suggestion.count ?? 0 > 0 ||
+                    message.suggestedReply()?.suggestion.count ?? 0 > 0 ||
+                    message.allButtons()?.suggestion.count ?? 0 > 0 ||
+                    currentMessageTime .int64Value >= lastSentMessageTime .int64Value {
+                    self.messageModels.append(message)
+                }
+            }
+        }
     }
 
     open func loadOpenGroupMessages() {
@@ -1722,7 +1765,7 @@ open class ALKConversationViewModel: NSObject, Localizable {
                 }
                 self.alMessageWrapper.getUpdatedMessageArray().insert(newMessages, at: 0)
                 self.alMessages.insert(mesg, at: 0)
-                self.messageModels.insert(mesg.messageModel, at: 0)
+                self.addMessageToMessageModel(messages: [mesg.messageModel])
             }
             self.delegate?.loadingFinished(error: nil)
         })